refactor: Move component directory
